# design_docs/openapi.v0.yaml
openapi: 3.0.3
info:
  title: Curious Now API (v0)
  version: "0.1"
servers:
  - url: /v1
paths:
...
  /auth/magic_link/start:
    post:
      summary: Stage 5 start email magic-link login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthMagicLinkStartRequest" }
      responses:
        "200":
          description: Sent (or silently accepted)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthMagicLinkStartResponse" }
  /auth/magic_link/verify:
    post:
      summary: Stage 5 verify magic-link token and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthMagicLinkVerifyRequest" }
      responses:
        "200":
          description: Verified; session created (cookie set)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthMagicLinkVerifyResponse" }
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/logout:
    post:
      summary: Stage 5 logout (revoke current session)
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LogoutResponse" }
  /user:
    get:
      summary: Stage 5 get the current user
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/prefs:
    get:
      summary: Stage 5 get user preferences and state
      security:
        - SessionAuth: []
      responses:
        "200":
          description: User prefs
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserPrefsResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    patch:
      summary: Stage 5 update user preferences
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserPrefsPatchRequest" }
      responses:
        "200":
          description: Updated prefs
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserPrefsResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/follows/topics/{topic_id}:
    post:
      summary: Stage 5 follow a topic
      security:
        - SessionAuth: []
      parameters:
        - in: path
          name: topic_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Followed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      summary: Stage 5 unfollow a topic
      security:
        - SessionAuth: []
      parameters:
        - in: path
          name: topic_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/blocks/sources/{source_id}:
    post:
      summary: Stage 5 block a source
      security:
        - SessionAuth: []
      parameters:
        - in: path
          name: source_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Blocked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      summary: Stage 5 unblock a source
      security:
        - SessionAuth: []
      parameters:
        - in: path
          name: source_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Unblocked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/saves/{cluster_id}:
    post:
      summary: Stage 5 save a cluster
      security:
        - SessionAuth: []
      parameters:
        - in: path
          name: cluster_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      summary: Stage 5 unsave a cluster
      security:
        - SessionAuth: []
      parameters:
        - in: path
          name: cluster_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Unsaved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/hides/{cluster_id}:
    post:
      summary: Stage 5 hide a cluster from all feeds
      security:
        - SessionAuth: []
      parameters:
        - in: path
          name: cluster_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Hidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      summary: Stage 5 unhide a cluster
      security:
        - SessionAuth: []
      parameters:
        - in: path
          name: cluster_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Unhidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/saves:
    get:
      summary: Stage 5 list saved clusters (reading list)
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Saved clusters
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserSavesResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/watches/clusters:
    get:
      summary: Stage 6 list watched clusters
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Watched clusters
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserWatchesResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/watches/clusters/{cluster_id}:
    post:
      summary: Stage 6 watch a cluster (story update alerts)
      security:
        - SessionAuth: []
      parameters:
        - in: path
          name: cluster_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Watched
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      summary: Stage 6 unwatch a cluster
      security:
        - SessionAuth: []
      parameters:
        - in: path
          name: cluster_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Unwatched
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /events:
    post:
      summary: Stage 5 engagement events (auth optional)
      security:
        - SessionAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventIn" }
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventsIngestResponse" }
...
components:
  securitySchemes:
    AdminTokenAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
    SessionAuth:
      type: apiKey
      in: cookie
      name: cn_session
  schemas:
    ContentType:
      type: string
      enum: [news, press_release, preprint, peer_reviewed, report]
...
    TopicChip:
      type: object
      required: [topic_id, name, score]
      properties:
        topic_id: { type: string, format: uuid }
        name: { type: string }
        score: { type: number }
    ClusterCard:
      type: object
      required: [cluster_id, canonical_title, updated_at, distinct_source_count]
      properties:
        cluster_id: { type: string, format: uuid }
        canonical_title: { type: string }
        updated_at: { type: string, format: date-time }
        distinct_source_count: { type: integer }
        top_topics:
          type: array
          items: { $ref: "#/components/schemas/TopicChip" }
        content_type_badges:
          type: array
          items: { $ref: "#/components/schemas/ContentType" }
        method_badges:
          type: array
          items: { type: string }
        takeaway: { type: string, nullable: true }
        confidence_band:
          type: string
          nullable: true
          enum: [early, growing, established]
        anti_hype_flags:
          type: array
          items: { type: string }
        is_saved: { type: boolean, nullable: true }
        is_watched: { type: boolean, nullable: true }
        why_in_feed: { $ref: "#/components/schemas/WhyInFeed" }
...
    WhyInFeed:
      type: object
      properties:
        reasons:
          type: array
          items: { type: string }
        matched_topic_ids:
          type: array
          items: { type: string, format: uuid }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    SimpleOkResponse:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [ok] }
    User:
      type: object
      required: [user_id, email, created_at]
      properties:
        user_id: { type: string, format: uuid }
        email: { type: string }
        created_at: { type: string, format: date-time }
    UserResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/User" }
    AuthMagicLinkStartRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    AuthMagicLinkStartResponse:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [sent] }
    AuthMagicLinkVerifyRequest:
      type: object
      required: [token]
      properties:
        token: { type: string, minLength: 20 }
    AuthMagicLinkVerifyResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/User" }
    LogoutResponse:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [ok] }
    UserPrefs:
      type: object
      required: [reading_mode_default, notification_settings]
      properties:
        reading_mode_default:
          type: string
          enum: [intuition, deep]
        followed_topic_ids:
          type: array
          items: { type: string, format: uuid }
        followed_entity_ids:
          type: array
          items: { type: string, format: uuid }
        blocked_source_ids:
          type: array
          items: { type: string, format: uuid }
        saved_cluster_ids:
          type: array
          items: { type: string, format: uuid }
        hidden_cluster_ids:
          type: array
          items: { type: string, format: uuid }
        notification_settings:
          type: object
          additionalProperties: true
    UserPrefsResponse:
      type: object
      required: [prefs]
      properties:
        prefs: { $ref: "#/components/schemas/UserPrefs" }
    UserPrefsPatchRequest:
      type: object
      properties:
        reading_mode_default:
          type: string
          enum: [intuition, deep]
        notification_settings:
          type: object
          additionalProperties: true
    SavedCluster:
      type: object
      required: [saved_at, cluster]
      properties:
        saved_at: { type: string, format: date-time }
        cluster: { $ref: "#/components/schemas/ClusterCard" }
    UserSavesResponse:
      type: object
      required: [saved]
      properties:
        saved:
          type: array
          items: { $ref: "#/components/schemas/SavedCluster" }
    WatchedCluster:
      type: object
      required: [watched_at, cluster]
      properties:
        watched_at: { type: string, format: date-time }
        cluster: { $ref: "#/components/schemas/ClusterCard" }
    UserWatchesResponse:
      type: object
      required: [watched]
      properties:
        watched:
          type: array
          items: { $ref: "#/components/schemas/WatchedCluster" }
...
    EventIn:
      type: object
      required: [event_type]
      properties:
        event_type:
          type: string
          enum:
            - open_cluster
            - click_item
            - save_cluster
            - unsave_cluster
            - hide_cluster
            - unhide_cluster
            - follow_topic
            - unfollow_topic
            - block_source
            - unblock_source
        client_id: { type: string, format: uuid, nullable: true }
        cluster_id: { type: string, format: uuid, nullable: true }
        item_id: { type: string, format: uuid, nullable: true }
        topic_id: { type: string, format: uuid, nullable: true }
        meta:
          type: object
          additionalProperties: true
    EventsIngestResponse:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [accepted] }
...